Adds a memory predicate moddule in contracts
