+++ src/goto-instrument/code_contracts.cpp
@@ -11,24 +11,21 @@ Date: February 2016
 /// \file
 /// Verify and use annotated invariants and pre/post-conditions
 
-#include "code_contracts.h"
+#include "contracts_memory_predicates.h"
 
 #include <algorithm>
 #include <map>
 
 #include <analyses/local_may_alias.h>
 
-#include <ansi-c/ansi_c_language.h>
 #include <ansi-c/c_expr.h>
-#include <ansi-c/expr2c.h>
+
+#include <goto-instrument/contracts_memory_predicates.h>
 
 #include <goto-programs/remove_skip.h>
 
-#include <linking/static_lifetime_init.h>
-
 #include <util/arith_tools.h>
 #include <util/c_types.h>
-#include <util/config.h>
 #include <util/expr_util.h>
 #include <util/fresh_symbol.h>
 #include <util/mathematical_expr.h>
@@ -36,72 +33,7 @@ Date: February 2016
 #include <util/message.h>
 #include <util/pointer_offset_size.h>
 #include <util/pointer_predicates.h>
-#include <util/prefix.h>
 #include <util/replace_symbol.h>
-
-bool return_value_visitort::found_return_value()
-{
-  return found;
-}
-
-void return_value_visitort::operator()(const exprt &exp)
-{
-  if(exp.id() != ID_symbol)
-    return;
-  const symbol_exprt &sym = to_symbol_expr(exp);
-  found |= sym.get_identifier() == CPROVER_PREFIX "return_value";
-}
-
-std::set<irep_idt> &functions_in_scope_visitort::function_calls()
-{
-  return function_set;
-}
-
-void functions_in_scope_visitort::operator()(const goto_programt &prog)
-{
-  forall_goto_program_instructions(ins, prog)
-  {
-    if(ins->is_function_call())
-    {
-      const code_function_callt &call = ins->get_function_call();
-
-      if(call.function().id() != ID_symbol)
-      {
-        log.error().source_location = call.find_source_location();
-        log.error() << "Function pointer used in function invoked by "
-                       "function contract: "
-                    << messaget::eom;
-        throw 0;
-      }
-      else
-      {
-        const irep_idt &fun_name =
-          to_symbol_expr(call.function()).get_identifier();
-        if(function_set.find(fun_name) == function_set.end())
-        {
-          function_set.insert(fun_name);
-          auto called_fun = goto_functions.function_map.find(fun_name);
-          if(called_fun == goto_functions.function_map.end())
-          {
-            log.warning() << "Could not find function '" << fun_name
-                          << "' in goto-program." << messaget::eom;
-            throw 0;
-          }
-          if(called_fun->second.body_available())
-          {
-            const goto_programt &program = called_fun->second.body;
-            (*this)(program);
-          }
-          else
-          {
-            log.warning() << "No body for function: " << fun_name
-                          << "invoked from function contract." << messaget::eom;
-          }
-        }
-      }
-    }
-  }
-}
 
 exprt get_size(const typet &type, const namespacet &ns, messaget &log)
 {
@@ -1763,30 +1695,16 @@ assigns_clauset::~assigns_clauset()
 
 assigns_clause_targett *assigns_clauset::add_target(exprt current_operation)
 {
-  if(current_operation.id() == ID_address_of)
-  {
-    assigns_clause_array_targett *array_target =
-      new assigns_clause_array_targett(
-        current_operation, parent, log, function_id);
-    targets.push_back(array_target);
-    return array_target;
-  }
-  else if(current_operation.type().id() == ID_struct_tag)
-  {
-    assigns_clause_struct_targett *struct_target =
-      new assigns_clause_struct_targett(
-        current_operation, parent, log, function_id);
-    targets.push_back(struct_target);
-    return struct_target;
-  }
-  else
-  {
-    assigns_clause_scalar_targett *scalar_target =
-      new assigns_clause_scalar_targett(
-        current_operation, parent, log, function_id);
-    targets.push_back(scalar_target);
-    return scalar_target;
-  }
+  assigns_clause_scalar_targett *scalar_target =
+    new assigns_clause_scalar_targett(
+      (current_operation.id() == ID_address_of)
+        ? to_index_expr(to_address_of_expr(current_operation).object()).array()
+        : current_operation,
+      parent,
+      log,
+      function_id);
+  targets.push_back(scalar_target);
+  return scalar_target;
 }
 
 assigns_clause_targett *
@@ -1867,6 +1785,7 @@ goto_programt assigns_clauset::havoc_code(
     goto_programt::targett z = tmp_z.add(goto_programt::make_skip(location));
 
     const auto &target_ptr = target->get_direct_pointer();
+
     if(to_address_of_expr(target_ptr).object().id() == ID_dereference)
     {
       // create the condition
@@ -1981,278 +1900,4 @@ exprt assigns_clauset::compatible_expression(
   }
 
   return result;
-}
-
-std::set<goto_programt::targett> &find_is_fresh_calls_visitort::is_fresh_calls()
-{
-  return function_set;
-}
-
-void find_is_fresh_calls_visitort::clear_set()
-{
-  function_set.clear();
-}
-
-void find_is_fresh_calls_visitort::operator()(goto_programt &prog)
-{
-  Forall_goto_program_instructions(ins, prog)
-  {
-    if(ins->is_function_call())
-    {
-      const code_function_callt &call = ins->get_function_call();
-
-      if(call.function().id() == ID_symbol)
-      {
-        const irep_idt &fun_name =
-          to_symbol_expr(call.function()).get_identifier();
-
-        if(fun_name == (CPROVER_PREFIX + std::string("is_fresh")))
-        {
-          function_set.insert(ins);
-        }
-      }
-    }
-  }
-}
-
-void is_fresh_baset::update_requires(goto_programt &requires)
-{
-  find_is_fresh_calls_visitort requires_visitor;
-  requires_visitor(requires);
-  for(auto it : requires_visitor.is_fresh_calls())
-  {
-    create_requires_fn_call(it);
-  }
-}
-
-void is_fresh_baset::update_ensures(goto_programt &ensures)
-{
-  find_is_fresh_calls_visitort ensures_visitor;
-  ensures_visitor(ensures);
-  for(auto it : ensures_visitor.is_fresh_calls())
-  {
-    create_ensures_fn_call(it);
-  }
-}
-
-//
-//
-// Code largely copied from model_argc_argv.cpp
-//
-//
-
-void is_fresh_baset::add_declarations(const std::string &decl_string)
-{
-  log.debug() << "Creating declarations: \n" << decl_string << "\n";
-
-  std::istringstream iss(decl_string);
-
-  ansi_c_languaget ansi_c_language;
-  ansi_c_language.set_message_handler(log.get_message_handler());
-  configt::ansi_ct::preprocessort pp = config.ansi_c.preprocessor;
-  config.ansi_c.preprocessor = configt::ansi_ct::preprocessort::NONE;
-  ansi_c_language.parse(iss, "");
-  config.ansi_c.preprocessor = pp;
-
-  symbol_tablet tmp_symbol_table;
-  ansi_c_language.typecheck(tmp_symbol_table, "<built-in-library>");
-  exprt value = nil_exprt();
-
-  goto_functionst tmp_functions;
-
-  // Add the new functions into the goto functions table.
-  parent.get_goto_functions().function_map[ensures_fn_name].copy_from(
-    tmp_functions.function_map[ensures_fn_name]);
-
-  parent.get_goto_functions().function_map[requires_fn_name].copy_from(
-    tmp_functions.function_map[requires_fn_name]);
-
-  for(const auto &symbol_pair : tmp_symbol_table.symbols)
-  {
-    if(
-      symbol_pair.first == memmap_name ||
-      symbol_pair.first == ensures_fn_name ||
-      symbol_pair.first == requires_fn_name || symbol_pair.first == "malloc")
-    {
-      this->parent.get_symbol_table().insert(symbol_pair.second);
-    }
-    // Parameters are stored as scoped names in the symbol table.
-    else if(
-      (has_prefix(
-         id2string(symbol_pair.first), id2string(ensures_fn_name) + "::") ||
-       has_prefix(
-         id2string(symbol_pair.first), id2string(requires_fn_name) + "::")) &&
-      parent.get_symbol_table().add(symbol_pair.second))
-    {
-      UNREACHABLE;
-    }
-  }
-
-  // We have to set the global memory map array to
-  // all zeros for this to work properly
-  const array_typet ty =
-    to_array_type(tmp_symbol_table.lookup_ref(memmap_name).type);
-  constant_exprt initial_value(irep_idt(dstringt("0")), ty.subtype());
-  array_of_exprt memmap_init(initial_value, ty);
-  goto_programt::instructiont a =
-    goto_programt::make_assignment(symbol_exprt(memmap_name, ty), memmap_init);
-
-  // insert the assignment into the initialize function.
-  auto called_func =
-    parent.get_goto_functions().function_map.find(INITIALIZE_FUNCTION);
-  goto_programt &body = called_func->second.body;
-  auto target = body.get_end_function();
-  body.insert_before(target, a);
-}
-
-void is_fresh_baset::update_fn_call(
-  goto_programt::targett &ins,
-  const std::string &fn_name,
-  bool add_address_of)
-{
-  const code_function_callt &const_call = ins->get_function_call();
-  code_function_callt call(
-    exprt(const_call.lhs()),
-    exprt(const_call.function()),
-    code_function_callt::argumentst(const_call.arguments()));
-
-  // adjusting the expression for the first argument, if required
-  if(add_address_of)
-  {
-    INVARIANT(call.arguments().size() > 0, "Function must have arguments");
-    call.arguments()[0] = address_of_exprt(call.arguments()[0]);
-  }
-
-  // fixing the function name.
-  to_symbol_expr(call.function()).set_identifier(fn_name);
-  log.debug() << "printing updated call expression: "
-              << expr2c(call, parent.get_namespace()) << "\n";
-
-  ins->set_function_call(call);
-}
-
-/* Declarations for contract enforcement */
-
-is_fresh_enforcet::is_fresh_enforcet(
-  code_contractst &_parent,
-  messaget _log,
-  irep_idt _fun_id)
-  : is_fresh_baset(_parent, _log, _fun_id)
-{
-  std::stringstream ssreq, ssensure, ssmemmap;
-  ssreq << CPROVER_PREFIX << fun_id << "_requires_is_fresh";
-  this->requires_fn_name = ssreq.str();
-
-  ssensure << CPROVER_PREFIX << fun_id << "_ensures_is_fresh";
-  this->ensures_fn_name = ssensure.str();
-
-  ssmemmap << CPROVER_PREFIX << fun_id << "_memory_map";
-  this->memmap_name = ssmemmap.str();
-}
-
-void is_fresh_enforcet::create_declarations()
-{
-  std::ostringstream oss;
-  std::string cprover_prefix(CPROVER_PREFIX);
-  oss << "static _Bool " << memmap_name
-      << "[" + cprover_prefix + "constant_infinity_uint]; \n"
-      << "\n"
-      << "_Bool " << requires_fn_name
-      << "(void **elem, " + cprover_prefix + "size_t size) { \n"
-      << "   *elem = malloc(size); \n"
-      << "   if (!*elem || " << memmap_name
-      << "[" + cprover_prefix + "POINTER_OBJECT(*elem)]) return 0; \n"
-      << "   " << memmap_name << "[" + cprover_prefix
-      << "POINTER_OBJECT(*elem)] = 1; \n"
-      << "   return 1; \n"
-      << "} \n"
-      << "\n"
-      << "_Bool " << ensures_fn_name
-      << "(void *elem, " + cprover_prefix + "size_t size) { \n"
-      << "   _Bool ok = (!" << memmap_name
-      << "[" + cprover_prefix + "POINTER_OBJECT(elem)] && "
-      << cprover_prefix + "r_ok(elem, size)); \n"
-      << "   " << memmap_name << "[" + cprover_prefix
-      << "POINTER_OBJECT(elem)] = 1; \n"
-      << "   return ok; \n"
-      << "}";
-
-  add_declarations(oss.str());
-}
-
-void is_fresh_enforcet::create_requires_fn_call(goto_programt::targett &ins)
-{
-  update_fn_call(ins, requires_fn_name, true);
-}
-
-void is_fresh_enforcet::create_ensures_fn_call(goto_programt::targett &ins)
-{
-  update_fn_call(ins, ensures_fn_name, false);
-}
-
-/* Declarations for contract replacement: note that there may be several
-   instances of the same function called in a particular context, so care must be taken
-   that the 'call' functions and global data structure are unique for each instance.
-   This is why we check that the symbols are unique for each such declaration.  */
-
-std::string unique_symbol(const symbol_tablet &tbl, const std::string &original)
-{
-  auto size = tbl.next_unused_suffix(original);
-  return original + std::to_string(size);
-}
-
-is_fresh_replacet::is_fresh_replacet(
-  code_contractst &_parent,
-  messaget _log,
-  irep_idt _fun_id)
-  : is_fresh_baset(_parent, _log, _fun_id)
-{
-  std::stringstream ssreq, ssensure, ssmemmap;
-  ssreq /* << CPROVER_PREFIX */ << fun_id << "_call_requires_is_fresh";
-  this->requires_fn_name =
-    unique_symbol(parent.get_symbol_table(), ssreq.str());
-
-  ssensure /* << CPROVER_PREFIX */ << fun_id << "_call_ensures_is_fresh";
-  this->ensures_fn_name =
-    unique_symbol(parent.get_symbol_table(), ssensure.str());
-
-  ssmemmap /* << CPROVER_PREFIX */ << fun_id << "_memory_map";
-  this->memmap_name = unique_symbol(parent.get_symbol_table(), ssmemmap.str());
-}
-
-void is_fresh_replacet::create_declarations()
-{
-  std::ostringstream oss;
-  std::string cprover_prefix(CPROVER_PREFIX);
-  oss << "static _Bool " << memmap_name
-      << "[" + cprover_prefix + "constant_infinity_uint]; \n"
-      << "\n"
-      << "static _Bool " << requires_fn_name
-      << "(void *elem, " + cprover_prefix + "size_t size) { \n"
-      << "  _Bool r_ok = " + cprover_prefix + "r_ok(elem, size); \n"
-      << "  if (" << memmap_name
-      << "[" + cprover_prefix + "POINTER_OBJECT(elem)]"
-      << " != 0 || !r_ok)  return 0; \n"
-      << "  " << memmap_name << "["
-      << cprover_prefix + "POINTER_OBJECT(elem)] = 1; \n"
-      << "  return 1; \n"
-      << "} \n"
-      << " \n"
-      << "_Bool " << ensures_fn_name
-      << "(void **elem, " + cprover_prefix + "size_t size) { \n"
-      << "  *elem = malloc(size); \n"
-      << "  return (*elem != 0); \n"
-      << "} \n";
-
-  add_declarations(oss.str());
-}
-
-void is_fresh_replacet::create_requires_fn_call(goto_programt::targett &ins)
-{
-  update_fn_call(ins, requires_fn_name, false);
-}
-
-void is_fresh_replacet::create_ensures_fn_call(goto_programt::targett &ins)
-{
-  update_fn_call(ins, ensures_fn_name, true);
 }
