+++ src/goto-instrument/code_contracts.h
@@ -31,6 +31,29 @@ Date: February 2016
 
 #include "loop_utils.h"
 
+#define FLAG_LOOP_CONTRACTS "apply-loop-contracts"
+#define HELP_LOOP_CONTRACTS                                                    \
+  " --apply-loop-contracts\n"                                                  \
+  "                              check and use loop contracts when provided\n"
+
+#define FLAG_REPLACE_CALL "replace-call-with-contract"
+#define HELP_REPLACE_CALL                                                      \
+  " --replace-call-with-contract <fun>\n"                                      \
+  "                              replace calls to fun with fun's contract\n"
+
+#define FLAG_REPLACE_ALL_CALLS "replace-all-calls-with-contracts"
+#define HELP_REPLACE_ALL_CALLS                                                 \
+  " --replace-all-calls-with-contracts\n"                                      \
+  "                              as above for all functions with a contract\n"
+
+#define FLAG_ENFORCE_CONTRACT "enforce-contract"
+#define HELP_ENFORCE_CONTRACT                                                  \
+  " --enforce-contract <fun>     wrap fun with an assertion of its contract\n"
+
+#define FLAG_ENFORCE_ALL_CONTRACTS "enforce-all-contracts"
+#define HELP_ENFORCE_ALL_CONTRACTS                                             \
+  " --enforce-all-contracts      as above for all functions with a contract\n"
+
 class assigns_clauset;
 class local_may_aliast;
 class replace_symbolt;
@@ -220,29 +243,6 @@ protected:
     const irep_idt &mode);
 };
 
-#define FLAG_LOOP_CONTRACTS "apply-loop-contracts"
-#define HELP_LOOP_CONTRACTS                                                    \
-  " --apply-loop-contracts\n"                                                  \
-  "                              check and use loop contracts when provided\n"
-
-#define FLAG_REPLACE_CALL "replace-call-with-contract"
-#define HELP_REPLACE_CALL                                                      \
-  " --replace-call-with-contract <fun>\n"                                      \
-  "                              replace calls to fun with fun's contract\n"
-
-#define FLAG_REPLACE_ALL_CALLS "replace-all-calls-with-contracts"
-#define HELP_REPLACE_ALL_CALLS                                                 \
-  " --replace-all-calls-with-contracts\n"                                      \
-  "                              as above for all functions with a contract\n"
-
-#define FLAG_ENFORCE_CONTRACT "enforce-contract"
-#define HELP_ENFORCE_CONTRACT                                                  \
-  " --enforce-contract <fun>     wrap fun with an assertion of its contract\n"
-
-#define FLAG_ENFORCE_ALL_CONTRACTS "enforce-all-contracts"
-#define HELP_ENFORCE_ALL_CONTRACTS                                             \
-  " --enforce-all-contracts      as above for all functions with a contract\n"
-
 /// \brief A base class for assigns clause targets
 class assigns_clause_targett
 {
@@ -405,146 +405,6 @@ protected:
   code_contractst &parent;
   const irep_idt function_id;
   messaget log;
-};
-
-class is_fresh_baset
-{
-public:
-  is_fresh_baset(
-    code_contractst &_parent,
-    messaget _log,
-    const irep_idt _fun_id)
-    : parent(_parent), log(_log), fun_id(_fun_id)
-  {
-  }
-
-  void update_requires(goto_programt &requires);
-  void update_ensures(goto_programt &ensures);
-
-  virtual void create_declarations() = 0;
-
-protected:
-  void add_declarations(const std::string &decl_string);
-  void update_fn_call(
-    goto_programt::targett &target,
-    const std::string &name,
-    bool add_address_of);
-
-  virtual void create_requires_fn_call(goto_programt::targett &target) = 0;
-  virtual void create_ensures_fn_call(goto_programt::targett &target) = 0;
-
-  code_contractst &parent;
-  messaget log;
-  irep_idt fun_id;
-
-  // written by the child classes.
-  std::string memmap_name;
-  std::string requires_fn_name;
-  std::string ensures_fn_name;
-};
-
-class is_fresh_enforcet : public is_fresh_baset
-{
-public:
-  is_fresh_enforcet(
-    code_contractst &_parent,
-    messaget _log,
-    const irep_idt _fun_id);
-
-  virtual void create_declarations();
-
-protected:
-  virtual void create_requires_fn_call(goto_programt::targett &target);
-  virtual void create_ensures_fn_call(goto_programt::targett &target);
-};
-
-class is_fresh_replacet : public is_fresh_baset
-{
-public:
-  is_fresh_replacet(
-    code_contractst &_parent,
-    messaget _log,
-    const irep_idt _fun_id);
-
-  virtual void create_declarations();
-
-protected:
-  virtual void create_requires_fn_call(goto_programt::targett &target);
-  virtual void create_ensures_fn_call(goto_programt::targett &target);
-};
-
-/// Predicate to be used with the exprt::visit() function.  It
-/// will return the set of function calls within a goto program.
-class find_is_fresh_calls_visitort
-{
-public:
-  find_is_fresh_calls_visitort()
-  {
-  }
-
-  // \brief return the set of functions invoked by
-  // the call graph of this program.
-  std::set<goto_programt::targett> &is_fresh_calls();
-  void clear_set();
-  void operator()(goto_programt &prog);
-
-protected:
-  std::set<goto_programt::targett> function_set;
-};
-
-/// Predicate to be used with the exprt::visit() function. The function
-/// found_return_value() will return `true` iff this predicate is called on an
-/// expr that contains `__CPROVER_return_value`.
-class return_value_visitort : public const_expr_visitort
-{
-public:
-  return_value_visitort() : const_expr_visitort(), found(false)
-  {
-  }
-
-  // \brief Has this object been passed to exprt::visit() on an exprt whose
-  //        descendants contain __CPROVER_return_value?
-  bool found_return_value();
-  void operator()(const exprt &exp) override;
-
-protected:
-  bool found;
-};
-
-/// Predicate to be used with the exprt::visit() function. The function
-/// found_return_value() will return `true` iff this predicate is called on an
-/// expr that contains `__CPROVER_return_value`.
-class functions_in_scope_visitort
-{
-public:
-  functions_in_scope_visitort(
-    const goto_functionst &goto_functions,
-    messaget &log)
-    : goto_functions(goto_functions), log(log)
-  {
-  }
-
-  // \brief return the set of functions invoked by
-  // the call graph of this program.
-  std::set<irep_idt> &function_calls();
-  void operator()(const goto_programt &prog);
-
-protected:
-  const goto_functionst &goto_functions;
-  messaget &log;
-  std::set<irep_idt> function_set;
-};
-
-class function_binding_visitort : const_expr_visitort
-{
-public:
-  function_binding_visitort() : const_expr_visitort()
-  {
-  }
-
-  void operator()(const exprt &exp) override
-  {
-  }
 };
 
 #endif // CPROVER_GOTO_INSTRUMENT_CODE_CONTRACTS_H
